Rename widget package to views.
