+++ app/src/main/java/com/chrisfry/socialq/userinterface/activities/ClientActivityKotlin.kt
@@ -15,7 +15,7 @@ import com.chrisfry.socialq.R
 import com.chrisfry.socialq.business.AppConstants
 import com.chrisfry.socialq.enums.RequestType
 import com.chrisfry.socialq.userinterface.adapters.BasicTrackListAdapter
-import com.chrisfry.socialq.userinterface.widgets.QueueItemDecoration
+import com.chrisfry.socialq.userinterface.views.QueueItemDecoration
 import com.spotify.sdk.android.authentication.AuthenticationClient
 import com.spotify.sdk.android.authentication.AuthenticationResponse
