+++ app/src/main/java/com/chrisfry/socialq/userinterface/activities/HostActivityKotlin.kt
@@ -20,7 +20,7 @@ import com.chrisfry.socialq.services.PlayQueueService
 import com.chrisfry.socialq.userinterface.adapters.HostTrackListAdapter
 import com.chrisfry.socialq.userinterface.adapters.IItemSelectionListener
 import com.chrisfry.socialq.userinterface.adapters.SelectablePlaylistAdapter
-import com.chrisfry.socialq.userinterface.widgets.QueueItemDecoration
+import com.chrisfry.socialq.userinterface.views.QueueItemDecoration
 import com.google.gson.JsonArray
 import com.spotify.sdk.android.authentication.AuthenticationClient
 import com.spotify.sdk.android.authentication.AuthenticationResponse
